backend/crm/requirements.txt
# Django
Django>=5.0.1,<6.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.1
django-filter>=24.0

# Database
psycopg[binary]>=3.1.0
dj-database-url>=2.1.0

# JWT validation
PyJWT>=2.8.0
python-jose[cryptography]>=3.3.0
cryptography>=42.0.0

# Configuration
python-dotenv>=1.0.0

# API docs
drf-yasg>=1.21.4

# Server
gunicorn>=21.2.0
whitenoise>=6.6.0

# Messaging
confluent-kafka>=2.3.0

# Redis (caching)
redis>=5.0.1

# HTTP client
httpx>=0.26.0

# Search (Elasticsearch)
elasticsearch>=8.11.0

# Validation
pydantic>=2.5.0
